server/lib/iolib/json.go
/*
 * Copyright (C) 2021 The Qing Project. All rights reserved.
 *
 * Use of this source code is governed by a license that can
 * be found in the LICENSE file.
 */

package iolib

import (
	"encoding/json"
	"os"
)

func ReadJSONFile(file string, v any) error {
	bytes, err := os.ReadFile(file)
	if err != nil {
		return err
	}
	return json.Unmarshal(bytes, v)
}

func WriteJSONFile(file string, v any) error {
	bytes, err := json.Marshal(v)
	if err != nil {
		return err
	}

	return WriteFile(file, bytes)
}

func StructToJSONMap(obj any) (map[string]any, error) {
	data, err := json.Marshal(obj)
	if err != nil {
		return nil, err
	}
	var m map[string]any
	err = json.Unmarshal(data, &m)
	if err != nil {
		return nil, err
	}
	return m, nil
}
